Implemented Matt's changes
